Fixing PSP builds, reenabling Dreamcast builds
1) .gitlab-ci.yml
    - reenabled Dreamcast build (without TLS until it makes it into KOS)
2) CMakeLists.txt
    - tried to move a bunch of PSP specific config variables to global
      scope to be consistent
3) gimbal_compiler.h
    - make GBL_THREAD_LOCAL basically a noop for Dreamcast (temporary)
      and PSP (maybe always)
4) gimbal_context.c
    - had to change %zu line number formatter to %u, because PSP doesn't
      have extended C99 formatters
5) gimbal_scanner.c
    - had to change GblScanner_raiseError() from constructing the error
      string + cursor location using %zu to %llu for line number and
column
6) test/CmakeLists.txt
    - had to link to a bunch of PSP debug libraries
7) gimbal_string_ref_test_suite.cpp
    - added a few more C++ binding test cases

+++ lib/source/utils/gimbal_scanner.c
@@ -148,9 +148,9 @@ GBL_EXPORT void GblScanner_raiseError(GblScanner* pSelf,
     GblStringBuffer_construct(pStrBuff, NULL, 0, bufferSize);
 
     GblStringBuffer_appendPrintf(pStrBuff,
-                                 "[line: %zu, column: %zu]: ",
-                                 GBL_SCANNER_CURSOR_(pSelf)->line   + 1,
-                                 GBL_SCANNER_CURSOR_(pSelf)->column + 1);
+                                 "[line: %lu, column: %lu]: ",
+                                 (unsigned long)GBL_SCANNER_CURSOR_(pSelf)->line   + 1,
+                                 (unsigned long)GBL_SCANNER_CURSOR_(pSelf)->column + 1);
 
     GblStringBuffer_appendVPrintf(pStrBuff,
                                   pFmt,
